hitcon-ctf-2018/oh-my-raddit/src/requirements.txt
pycrypto==2.6.1
web.py==0.38
